requirements.txt
pytorch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2
numpy
PyYAML
imageio
imageio-ffmpeg
matplotlib
opencv-python
scikit-image
tqdm
h5py
progressbar
psutil
ninja
gdown
zipfile36
argparse
torchmetrics
